python/requirements.txt
# Core dependencies
youtube-transcript-api>=0.6.1
whisper-openai>=1.0.0
torch>=2.0.0
numpy>=1.24.0

# Audio processing
ffmpeg-python>=0.2.0

# Utility packages
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional but recommended for better performance
setuptools>=68.0.0
wheel>=0.40.0